Update rand_data.lua
The output (y) is given as 1 or 2 because the loss function has an issue of only working on unit length non zero values. I'm not exactly sure about this though. I was stuck for a few hours trying to figure out what was wrong. The code kept giving bugs when run with zeros or multiple dimensions. 

The docs say that only 1D vectors as targets work for the loss function.